(single-focus-trap) Remove all the logic around MutationObserver API
It can observe the style of an element, but it doesn't account for computed styles.
Since `getComputedStyles` returns a live `CSSStyleDeclaration`  I could work my way around the issue by observing the computed styles through a `Proxy` object.
Another solution would be to use an `IntersectionObserver` to spot when an element appear/disappear from the document, though this approach would fail in case an element should not be focusable even if it takes up space in the document (and only if this mutatiion was about CSS styles), eg: "visibility: hidden".
But I think that in the end the performance gains of these approaches are not worth their implementatioon and mainteinance.

About the performance gains:
Without observers the trap has to be updated at every `Tab` key press.
With observers a trap update is scheduled only after a particular mutation happened and then when a `Tab` key press occurs, the trap is updated only if an update was scheduled.
In practice, with observers the trap would update almost never thanks to an optimization that is possible whenever there are no elements with a positive tab index in the trap (very likely).

Another thing to notice is that:
Without observers the focus trap's footprint on performance would be basically inhexistent for users that do not use the keyboard to navigate through the page.
While using observers means having them up and running as long as the trap is active. Though I believe the footprint on performance would still verge on zero, since their callback wouldn't do any expensive computation.

Updating a trap means:
- Querying all the possible focusable elements in the `roots`
- Iteratively picking those that are more likely to be the "kingpins" of each root
- Running checks on them to ensure thay are actually tabbable
- Building the arrays of kingpins

I'm gonna tag this commit, just in case I reconsider my position on whether the performance gains are worth the implementation.