(single-focus-trap) Perf e2e tests
When skipping to check that elements outside the trap pass the focus to the right elements inside the trap, it's possible to call `getTabCycle` just once, with a `len` equal to the length of the expected cycle.
This is much faster than calling `getTabCycle` with a `len === 2`  for each element in `@possibleTabbables`.