Remove Vercel GitHub App from project
It used to automatically deploy the demo after pushes to the main branch.
The behaviour of the deployment was being controlled by a vercel.deploy.js script, which is being removed.

The deployment is going to be controlled via GitHub Actions, allowing for more freedom of movement.
(The aforementioned script could only access the last 10 commits of the repo)