Settle for manual deployment of demo app
I wanted to deploy based on the presence of a `DEPLOY_SIGNATURE` in one of the pushed commit messages.

I could do so in a script for the "Ignored Build Step": https://vercel.com/guides/how-do-i-use-the-ignored-build-step-field-on-vercel
In which a `VERCEL_GIT_PREVIOUS_SHA` environment variable is aveilable: https://github.com/vercel/vercel/discussions/7251#discussioncomment-4731588
But the ignored build step can only access the last 10 commits of the repo.

So I opted for the Bild Output Api: https://vercel.com/blog/build-output-api
Here I can access all the git history needed, build the project and trigger a deploy on vercel.
However there is no `VERCEL_GIT_PREVIOUS_SHA` variable available and knowing which commit was the last deployed one would be tricky:
(I considered using the "git-auto-commit" action https://github.com/marketplace/actions/git-auto-commit
as well as implementing a moving `LAST DEPLOY` tag set by a git hook)

The fact is that I rarely want to deploy the demo, and a manual deploy is a quick and easy thing to get done via the Vercel CLI: https://vercel.com/docs/cli/deploy
Threfore I believe that deploying manually is the right approach for my usecase.