(demo/src/components/playground) Rename demo-element-controls to demo-button-controls
It was named "element" because I wanted to stay open to the possibility of using elements other than buttons.
Only now I can see that making the current code less descriptive for the sake of a feature I don't even know if and when will be implemented, is not a good idea.
I prefer having as good as possible namings at all times and making the necessary changes only if and when required.