(focus-trap) Refactor e2e tests
Rather than using the UI of the "playground" in the demo to build a focus trap, now tests use `searchParams` that allow for building traps with arbitrary `trapArgs`.
One limitation of this approach is that the `trapArg` need to be URI-encodable and JSON-serialisable.
However testing paramenters such as functions may be possible using a flag and a bit of coordination between the demo and the tests.