(focus-trap/src) Move sorting of `roots` from normalise.ts to destination.ts
Even though it makes sense to sort the roots when normalising them, they are consumed only by destination.ts, which in turn relies on the `roots` being sorted to find the right destination.
So it makes more sense to have the sorting in destination.ts, making it kind of self contained.