(focus-trap/src) Remove unnecessary `isEsc` flag
It was meant to distinguish between a "DEMOLISH" action and an `Esc` key press, so that errors wouldn't be thrown repeatedly when repeatedly pressing `Esc`.
However after the first `Esc`, there are no more listeners attached to the DOM, so further `Esc` are ignored by the trap.
There are no listeners attached to the DOM