(focus-trap) Move "true-myth" from "devDependencies" to "dependencies" and remove it from the bundle
I think that having it in "devDependencies" was just fine, as it's just an internal tool whose API is not exposed to the user of the library.
However, it does make sense to use "dependencies" to differentiate between code needed in development and code needed at runtime (see this comment: https://github.com/facebook/create-react-app/issues/1764#issuecomment-285082921).
I'll leave the burden of bundling 'true-myth' to the user of the library.