(focus-tra/cypress/e2e/trap-config/returnFocus.cy.ts) Fix test for `returnFocus` set to `false`
The test was obviously passing because of the "bug" documented within the test itself (at least I think it's a bug, still need to open an issue).
The final statement is checking whether an element is focused, but because of the supposed bug, it was checking whether `document.activeElement` is focused, and obviously it is.