(focus-trap/src/destination.ts) Improve performance of `nextPositiveTabbable`
The function was unnecessarily inefficient in case `origin` didn't belong to the trap.
Replaced a `push` + `sort`, probably O(n*log(n)), with a `splice` at the right `findIndex`, O(n).

It's not going to make any noticeable difference and the sorting approach is more concide and easier to follow.
So I'm gonna revert this commit right away.

I wrote it just so that I will remember and stop being tempted to fix it.